aegis-service/workspace/aegis_workspace/agents/travel_planner/requirements.txt
aegis-sdk
python-dotenv
aiohttp
beautifulsoup4
serpapi
googlemaps